feat: выделены общие Enums
