+++ TeamPlayerProfiles/Common/Models/Enums/MainStat.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Common.Models.Enums
+{
+    public enum MainStat
+    {
+        Agility,
+        Intelligence,
+        Strength,
+        Universal
+    }
+}
