+++ TeamPlayerProfiles/Common/Models/Enums/Position.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Common.Models.Enums
+{
+    public enum Position
+    {
+        Carry = 1,
+        Midlane = 2,
+        Offlane = 3,
+        Roamer = 4,
+        Support = 5,
+    }
+}
