+++ TeamPlayerProfiles/DataAccess/Context/EntitiesConfigurations/PositionConfiguration.cs
@@ -1,7 +1,7 @@
 ﻿using DataAccess.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Enums = DataAccess.Entities.Enums;
+using EnumPosition = Common.Models.Enums.Position;
 
 namespace DataAccess.Context.EntitiesConfigurations
 {
@@ -11,18 +11,18 @@ namespace DataAccess.Context.EntitiesConfigurations
         {
             builder.Property(e => e.Name)
                 .IsRequired()
-                .HasConversion(e => e.ToString(), e => Enum.Parse<Enums.Position>(e));
+                .HasConversion(e => e.ToString(), e => Enum.Parse<EnumPosition>(e));
 
             SeedData(builder);
         }
 
         private void SeedData(EntityTypeBuilder<Position> builder)
         {
-            builder.HasData(new Position { Id = 1, Name = Enums.Position.Carry });
-            builder.HasData(new Position { Id = 2, Name = Enums.Position.Midlane });
-            builder.HasData(new Position { Id = 3, Name = Enums.Position.Offlane });
-            builder.HasData(new Position { Id = 4, Name = Enums.Position.Roamer });
-            builder.HasData(new Position { Id = 5, Name = Enums.Position.Support });
+            builder.HasData(new Position { Id = 1, Name = EnumPosition.Carry });
+            builder.HasData(new Position { Id = 2, Name = EnumPosition.Midlane });
+            builder.HasData(new Position { Id = 3, Name = EnumPosition.Offlane });
+            builder.HasData(new Position { Id = 4, Name = EnumPosition.Roamer });
+            builder.HasData(new Position { Id = 5, Name = EnumPosition.Support });
         }
     }
 }
