+++ TeamPlayerProfiles/DataAccess/DataAccess.csproj
@@ -26,4 +26,8 @@
 	  <Folder Include="Repositories\Migrations\" />
 	</ItemGroup>
 
+	<ItemGroup>
+	  <ProjectReference Include="..\Common\Common.csproj" />
+	</ItemGroup>
+
 </Project>
