--- TeamPlayerProfiles/DataAccess/Entities/Enums/MainStat.cs
@@ -1,10 +1,0 @@
-﻿namespace DataAccess.Entities.Enums
-{
-    public enum MainStat
-    {
-        Agility,
-        Intelligence,
-        Strength,
-        Universal
-    }
-}
