--- TeamPlayerProfiles/DataAccess/Entities/Enums/Position.cs
@@ -1,11 +1,0 @@
-﻿namespace DataAccess.Entities.Enums
-{
-    public enum Position
-    {
-        Carry = 1,
-        Midlane = 2,
-        Offlane = 3,
-        Roamer = 4,
-        Support = 5,
-    }
-}
