+++ TeamPlayerProfiles/Service/Contracts/Player/PlayerDto.cs
@@ -1,4 +1,5 @@
 ﻿using Service.Contracts.Hero;
+using EnumPosition = Common.Models.Enums.Position;
 
 namespace Service.Contracts.Player
 {
@@ -14,7 +15,7 @@ namespace Service.Contracts.Player
 
         public bool Displayed { get; set; }
 
-        public string Position { get; set; }
+        public EnumPosition Position { get; set; }
 
         public ICollection<HeroDto> Heroes { get; set; }
