+++ TeamPlayerProfiles/Service/Contracts/Team/PlayerInTeamDto.cs
@@ -1,8 +1,10 @@
-﻿namespace Service.Contracts.Team
+﻿using EnumPosition = Common.Models.Enums.Position;
+
+namespace Service.Contracts.Team
 {
     public class PlayerInTeamDto
     {
-        public string Position { get; set; }
+        public EnumPosition Position { get; set; }
 
         public Guid PlayerId { get; set; }
     }
