+++ TeamPlayerProfiles/Service/Mappings/PlayerMappingProfile.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using DataAccess.Entities;
 using Service.Contracts.Player;
-using Enums = DataAccess.Entities.Enums;
+using EnumPosition = Common.Models.Enums.Position;
 
 namespace Service.Mapping
 {
@@ -12,17 +12,14 @@ namespace Service.Mapping
             CreateMap<Player, PlayerDto>();
 
             CreateMap<CreatePlayerDto, Player>()
-                .ForMember(d => d.Position, m => m.MapFrom(src => Enum.Parse<Enums.Position>(src.Position)))
-                .ForMember(d => d.PositionId, m => m.MapFrom(src => (int)Enum.Parse<Enums.Position>(src.Position)))
+                .ForMember(d => d.PositionId, m => m.MapFrom(src => (int)src.Position))
                 .ForMember(d => d.Displayed, m => m.Ignore())
                 .ForMember(d => d.Teams, m => m.Ignore())
                 .ForMember(d => d.TeamPlayers, m => m.Ignore());
 
             CreateMap<UpdatePlayerDto, Player>()
                 .ForMember(d => d.UserId, m => m.Ignore())
-                .ForMember(d => d.Position, m => m.MapFrom(src => Enum.Parse<Enums.Position>(src.Position)))
-                .ForMember(d => d.PositionId, m => m.MapFrom(src => (int)Enum.Parse<Enums.Position>(src.Position)))
-                .ForMember(d => d.Displayed, m => m.Ignore())
+                .ForMember(d => d.PositionId, m => m.MapFrom(src => (int)src.Position))
                 .ForMember(d => d.Teams, m => m.Ignore())
                 .ForMember(d => d.TeamPlayers, m => m.Ignore());
         }
